Edits on preprocessing side of pipeline
Here multiple options for applying and dealing with preprocessing are implemented. Now we are able to apply z-score and robust min-max normalization, resampling on the fly if necessary, bias correction on the fly if necessary and otsu thresholding on the fly if necessary. Furthermore, the np.nonzero function in the niftidataset.py file has been replaced by a cv2 implementation to allow for a speedup when doing normalization.